requirements.txt
click
twisted == 12.2.*
ldaptor
six
pbr
PyYAML
dogpile.cache
